Fix compiler warnings about dead code.
There is a setting that was meant to control whether XCode's "clean" action would remove libraries.  But since build scripts aren't invoked on XCode clean, there's no receiver code which uses this setting.  So the setting is dead code; this suppresses the warning about that but leaves the setting in place.